+++ packages/defiscan-frontend/src/pages/review/views/explorer/AdminsTab.tsx
@@ -4,12 +4,8 @@ import { AddressDisplay } from '../../../../components/AddressDisplay'
 import { UsdValue } from '../../../../components/UsdValue'
 import { formatUsdValue } from '../../../../utils/format'
 import { getHumanAdmins } from '../../../../utils/admins'
-import {
-  displayMitigationValue,
-  type CompiledReview,
-  type CompiledAdmin,
-  type Mitigation,
-} from '../../../../types'
+import { type CompiledReview, type CompiledAdmin } from '../../../../types'
+import { MitigationBadge } from '../../../../components/MitigationBadge'
 
 interface AdminsTabProps {
   review: CompiledReview
@@ -277,7 +273,7 @@ function ExpandedFunctions({ admin }: { admin: CompiledAdmin }) {
               <th className="text-left pb-1 font-medium">Contract</th>
               <th className="text-left pb-1 font-medium">Function</th>
               <th className="text-left pb-1 font-medium">Mitigations</th>
-              <th className="text-right pb-1 font-medium">Direct $</th>
+              <th className="text-right pb-1 font-medium">TVL</th>
               <th className="text-right pb-1 font-medium">
                 Reachable Contracts
               </th>
@@ -396,70 +392,3 @@ function SortHeader({
   )
 }
 
-function formatDelayLabel(seconds: number): string {
-  if (seconds >= 86400) {
-    const days = seconds / 86400
-    const d = days === Math.floor(days) ? `${days}` : `${days.toFixed(1)}`
-    return `${d}-day Delay`
-  }
-  if (seconds >= 3600) {
-    const hours = seconds / 3600
-    const h = hours === Math.floor(hours) ? `${hours}` : `${hours.toFixed(1)}`
-    return `${h}-hour Delay`
-  }
-  if (seconds >= 60) {
-    const minutes = seconds / 60
-    const m = minutes === Math.floor(minutes) ? `${minutes}` : `${minutes.toFixed(1)}`
-    return `${m}-min Delay`
-  }
-  return `${seconds}s Delay`
-}
-
-function MitigationBadge({ mitigation: m }: { mitigation: Mitigation }) {
-  const colorClass =
-    m.type === 'delay'
-      ? 'bg-cyan-50 text-cyan-700 border-cyan-200'
-      : m.type === 'valueRange'
-        ? 'bg-indigo-50 text-indigo-700 border-indigo-200'
-        : m.type === 'relativeValue'
-          ? 'bg-amber-50 text-amber-700 border-amber-200'
-          : 'bg-gray-50 text-gray-600 border-gray-200'
-
-  let label: string
-  let tooltip: string
-  if (m.type === 'delay') {
-    label =
-      m.delaySeconds !== undefined
-        ? formatDelayLabel(m.delaySeconds)
-        : 'Delay'
-    tooltip = m.description
-  } else if (m.type === 'valueRange') {
-    const parts: string[] = []
-    if (m.valueRange?.min !== undefined)
-      parts.push(displayMitigationValue(m.valueRange.min))
-    if (m.valueRange?.max !== undefined)
-      parts.push(displayMitigationValue(m.valueRange.max))
-    const unit = m.valueRange?.unit ? ` ${m.valueRange.unit}` : ''
-    label = `Range: ${parts.join(' to ')}${unit}`
-    tooltip = m.description || label
-  } else if (m.type === 'relativeValue') {
-    label = 'Relative'
-    tooltip = `Max change: ${m.relativeValue?.maxChangePercent !== undefined ? displayMitigationValue(m.relativeValue.maxChangePercent) : '?'}%`
-    if (m.description) tooltip += ` — ${m.description}`
-  } else {
-    label =
-      m.description.length > 20
-        ? m.description.slice(0, 20) + '…'
-        : m.description
-    tooltip = m.description
-  }
-
-  return (
-    <span
-      className={`inline-block rounded-full border px-1.5 py-0 text-[10px] font-medium leading-4 ${colorClass}`}
-      title={tooltip}
-    >
-      {label}
-    </span>
-  )
-}
