+++ packages/defiscan-frontend/src/pages/review/views/explorer/DepsTab.tsx
@@ -109,7 +109,7 @@ export function DepsTab({ review }: DepsTabProps) {
         )}
         {totalFundsAtRisk > 0 && (
           <span className="text-text-secondary">
-            TVL exposed:{' '}
+            TVL{' '}
             <UsdValue
               value={totalFundsAtRisk}
               variant="capital"
@@ -160,7 +160,7 @@ export function DepsTab({ review }: DepsTabProps) {
               </th>
               <SortHeader
                 field="fundsAtRisk"
-                label="TVL Exposed"
+                label="TVL"
                 current={sortField}
                 dir={sortDir}
                 onClick={handleSort}
@@ -357,7 +357,7 @@ function FunctionList({
         <tr className="text-text-muted">
           <th className="text-left pb-1 font-medium">Contract</th>
           <th className="text-left pb-1 font-medium">Function</th>
-          <th className="text-right pb-1 font-medium">TVL Exposed</th>
+          <th className="text-right pb-1 font-medium">TVL</th>
         </tr>
       </thead>
       <tbody>
