+++ packages/defiscan-frontend/src/utils/narrative.ts
@@ -130,7 +130,7 @@ export function getKeyFindings(review: CompiledReview): KeyFinding[] {
     const maxCapital = Math.max(...eoas.map((e) => e.totalDirectCapital))
     findings.push({
       type: 'critical',
-      title: `${eoas.length} EOA${eoas.length > 1 ? 's' : ''} with admin access`,
+      title: `${eoas.length} EOA${eoas.length > 1 ? 's' : ''} with permissioned access`,
       detail: `Externally owned account${eoas.length > 1 ? 's' : ''} can execute critical functions affecting up to ${formatUsdValue(maxCapital)} in TVL.`,
     })
   }
